chore: update dependencies and update & restructure tests to match [publish]
