+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
 	globals: {
 		'ts-jest': {
-			tsConfig: 'tsconfig.json',
+			tsconfig: 'tsconfig.json',
 			diagnostics: {
 				ignoreCodes: []
 			}
