+++ src/__tests__/MosConnection.spec.ts
@@ -1,199 +1,29 @@
-import '../__mocks__/time'
+import { clearMocks,
+	decode,
+	delay,
+	encode,
+	getXMLReply,
+	initMosConnection,
+	setupMocks
+} from './lib'
 import {
-	IMOSAck,
 	MosConnection,
 	MosDevice,
 	IMOSConnectionStatus,
-	IMOSObject,
-	IMOSItem,
-	IMOSItemAction,
-	IMOSItemStatus,
-	IMOSObjectList,
-	IMOSRequestObjectList,
-	IMOSROAck,
-	IMOSROAction,
-	IMOSROFullStory,
-	IMOSROReadyToAir,
-	IMOSROStory,
-	IMOSRunningOrder,
-	IMOSRunningOrderBase,
-	IMOSRunningOrderStatus,
-	IMOSListSearchableSchema,
-	IMOSStoryAction,
-	IMOSStoryStatus,
-	MosString128,
-	IMOSAckStatus,
-	IMOSObjectType,
-	IMOSObjectStatus,
-	IMOSObjectAirStatus,
-	IProfiles,
-	IMOSListMachInfo
-} from '../'
-
+	IMOSObject} from '../'
 import { ConnectionConfig } from '../config/connectionConfig'
-
-// @ts-ignore Socket is never read
-import { Socket, Server } from 'net'
-
 import { SocketMock } from '../__mocks__/socket'
 import { ServerMock } from '../__mocks__/server'
-
 import { xmlData, xmlApiData } from '../__mocks__/testData'
-import { xml2js } from 'xml-js'
 
-import * as Utils from '../utils/Utils'
-
-import * as iconv from 'iconv-lite'
-iconv.encodingExists('utf16-be')
-
-// breaks net.Server, disabled for now
-jest.mock('net')
-
-let setTimeoutOrg = setTimeout
-let delay = (ms: number) => {
-	return new Promise((resolve) => {
-		setTimeoutOrg(resolve, ms)
-	})
-}
-
-async function getMosConnection (profiles: IProfiles, strict: boolean): Promise<MosConnection> {
-	ServerMock.mockClear()
-
-	let mos = new MosConnection({
-		mosID: 'our.mos.id',
-		acceptsConnections: true,
-		profiles: profiles,
-		strict: strict,
-		debug: false
-	})
-	await mos.init()
-
-	expect(ServerMock.instances).toHaveLength(3)
-
-	return Promise.resolve(mos)
-}
-async function getMosDevice (mos: MosConnection): Promise<MosDevice> {
-	SocketMock.mockClear()
-
-	let device = await mos.connect({
-		primary: {
-			id: 'their.mos.id',
-			host: '127.0.0.1',
-			timeout: 200
-		}
-	})
-
-	await delay(10) // to allow for async timers & events to triggered
-
-	return Promise.resolve(device)
-}
-let fakeIncomingMessageMessageId = 1632
-function fakeIncomingMessage (socketMockLower: SocketMock, message: string, ourMosId?: string, theirMosId?: string): Promise<number> {
-	fakeIncomingMessageMessageId++
-	let fullMessage = getXMLReply(fakeIncomingMessageMessageId, message, ourMosId, theirMosId)
-	socketMockLower.mockReceiveMessage(encode(fullMessage))
-
-	return Promise.resolve(fakeIncomingMessageMessageId)
-}
-function getXMLReply (messageId: string | number, content: string, ourMosId?: string, theirMosId?: string): string {
-	return '<mos>' +
-		'<mosID>' + (ourMosId || 'our.mos.id') + '</mosID>' +
-		'<ncsID>' + (theirMosId || 'their.mos.id') + '</ncsID>' +
-		'<messageID>' + messageId + '</messageID>' +
-		content +
-		'</mos>\r\n'
-}
-function decode (data: Buffer): string {
-	return iconv.decode(data, 'utf16-be')
-}
-function encode (str: string) {
-	return iconv.encode(str, 'utf16-be')
-}
-async function checkReplyToServer (socket: SocketMock, messageId: number, replyString: string) {
-	// check reply to server:
-	await socket.mockWaitForSentMessages()
-
-	expect(socket.mockSentMessage).toHaveBeenCalledTimes(1)
-	// @ts-ignore mock
-	let reply = decode(socket.mockSentMessage.mock.calls[0][0])
-	expect(reply).toContain('<messageID>' + messageId + '</messageID>')
-	expect(reply).toContain(replyString)
-	expect(reply).toMatchSnapshot(reply)
-}
-function checkMessageSnapshot (msg: string) {
-	expect(
-		msg
-		.replace(/<messageID>\d+<\/messageID>/, '<messageID>xx</messageID>')
-		.replace(/<time>[^<>]+<\/time>/, '<time>xx</time>')
-	).toMatchSnapshot()
-
-}
-function checkAckSnapshot (ack: IMOSAck | IMOSROAck) {
-	const ack2: any = {
-		...ack
-	}
-	if (ack2.mos) {
-		ack2.mos = {
-			...ack2.mos,
-			messageID: 999
-		}
-	}
-	expect(ack2).toMatchSnapshot()
-
-}
-function doBeforeAll () {
-	let socketMockLower = SocketMock.instances.find(s => s.connectedPort === 10540) as SocketMock
-	let socketMockUpper = SocketMock.instances.find(s => s.connectedPort === 10541) as SocketMock
-	let socketMockQuery = SocketMock.instances.find(s => s.connectedPort === 10542) as SocketMock
-
-	expect(socketMockLower).toBeTruthy()
-	expect(socketMockUpper).toBeTruthy()
-	expect(socketMockQuery).toBeTruthy()
-
-	expect(ServerMock.instances).toHaveLength(3)
-
-	let serverMockLower = ServerMock.instances[0]
-	let serverMockUpper = ServerMock.instances[2]
-	let serverMockQuery = ServerMock.instances[1]
-	expect(serverMockLower).toBeTruthy()
-	expect(serverMockUpper).toBeTruthy()
-	expect(serverMockQuery).toBeTruthy()
-
-	// Pretend a server connects to us:
-	let serverSocketMockLower = serverMockLower.mockNewConnection()
-	let serverSocketMockUpper = serverMockUpper.mockNewConnection()
-	let serverSocketMockQuery = serverMockQuery.mockNewConnection()
-
-	socketMockLower.name = 'lower'
-	socketMockUpper.name = 'upper'
-	socketMockQuery.name = 'query'
-
-	serverSocketMockLower.name = 'serverLower'
-	serverSocketMockUpper.name = 'serverUpper'
-	serverSocketMockQuery.name = 'serverQuery'
-
-	return {
-		socketMockLower,
-		socketMockUpper,
-		socketMockQuery,
-		serverMockLower,
-		serverMockUpper,
-		serverMockQuery,
-		serverSocketMockLower,
-		serverSocketMockUpper,
-		serverSocketMockQuery
-	}
-}
+// @ts-ignore imports are unused
+import { Socket } from 'net'
 
 beforeAll(() => {
-	// Mock tcp connection
-	// @ts-ignore Replace Socket with the mocked varaint:
-	Socket = SocketMock
-	// @ts-ignore Replace Server with the mocked varaint:
-	Server = ServerMock
+	setupMocks()
 })
 beforeEach(() => {
-	SocketMock.mockClear()
+	clearMocks()
 })
 describe('MosDevice: General', () => {
 	test('Test the Socket mock', async () => {
@@ -260,6 +90,7 @@ describe('MosDevice: General', () => {
 			'7': false
 		})
 
+		await mos.dispose()
 	})
 	test('Incoming connections', async () => {
 		let mos = new MosConnection({
@@ -271,7 +102,7 @@ describe('MosDevice: General', () => {
 			}
 		})
 		expect(mos.acceptsConnections).toBe(true)
-		await mos.init()
+		await initMosConnection(mos)
 		expect(mos.isListening).toBe(true)
 		expect(SocketMock.instances).toHaveLength(0)
 
@@ -288,7 +119,7 @@ describe('MosDevice: General', () => {
 			}
 		})
 		expect(mos.acceptsConnections).toBe(true)
-		await mos.init()
+		await initMosConnection(mos)
 		expect(mos.isListening).toBe(true)
 
 		let mosDevice = await mos.connect({
@@ -327,7 +158,7 @@ describe('MosDevice: General', () => {
 			}
 		})
 		expect(mos.acceptsConnections).toBe(true)
-		await mos.init()
+		await initMosConnection(mos)
 		expect(mos.isListening).toBe(true)
 
 		let mosDevice = await mos.connect({
@@ -379,6 +210,7 @@ describe('MosDevice: General', () => {
 		returnedObj = await mosDevice.sendRequestMOSObject(xmlApiData.mosObj.ID!)
 		expect(returnedObj).toBeTruthy()
 
+		await mos.dispose()
 	})
 	test('init and connectionStatusChanged', async () => {
 		SocketMock.mockClear()
@@ -392,7 +224,7 @@ describe('MosDevice: General', () => {
 				'1': true
 			}
 		}))
-		await mosConnection.init()
+		await initMosConnection(mosConnection)
 
 		let mosDevice = await mosConnection.connect({
 			primary: {
@@ -434,6 +266,8 @@ describe('MosDevice: General', () => {
 
 		// @todo: add timeout test
 		// mock cause timeout
+
+		await mosConnection.dispose()
 	})
 	test('buddy failover', async () => {
 
@@ -445,7 +279,7 @@ describe('MosDevice: General', () => {
 				'1': true
 			}
 		})
-		await mos.init()
+		await initMosConnection(mos)
 		let mosDevice: MosDevice = await mos.connect({
 			primary: {
 				id: 'mockServer',
@@ -494,7 +328,7 @@ describe('MosDevice: General', () => {
 		connMocks[1].setReplyToHeartBeat(false)
 		connMocks[2].setReplyToHeartBeat(true)
 
-		await delay(500) // to allow for timeout:
+		await delay(600) // to allow for timeout:
 		expect(mosDevice.getConnectionStatus()).toMatchObject({
 			PrimaryConnected: false,
 			PrimaryStatus: '', // if not connected this will contain human-readable error-message
@@ -510,1266 +344,7 @@ describe('MosDevice: General', () => {
 
 		expect(connMocks[1].destroy).toHaveBeenCalledTimes(1)
 		expect(connMocks[2].destroy).toHaveBeenCalledTimes(1)
-	})
-})
 
-describe('Profile 0', () => {
-	let mosDevice: MosDevice
-	let mosConnection: MosConnection
-
-	let socketMockLower: SocketMock
-	let socketMockUpper: SocketMock
-	let socketMockQuery: SocketMock
-
-	let serverSocketMockLower: SocketMock
-	let serverSocketMockUpper: SocketMock
-	let serverSocketMockQuery: SocketMock
-
-	let onRequestMachineInfo: jest.Mock<any, any>
-	let onRequestMOSObject: jest.Mock<any, any>
-	let onRequestAllMOSObjects: jest.Mock<any, any>
-
-	beforeAll(async () => {
-
-		mosConnection = await getMosConnection({
-			'0': true,
-			'1': true // Must support at least one other profile
-		}, true)
-		mosDevice = await getMosDevice(mosConnection)
-
-		// Profile 0:
-		onRequestMachineInfo = jest.fn(() => {
-			return Promise.resolve(xmlApiData.machineInfo)
-		})
-		mosDevice.onRequestMachineInfo((): Promise<IMOSListMachInfo> => {
-			return onRequestMachineInfo()
-		})
-		// Profile 1:
-		onRequestMOSObject = jest.fn(() => {
-			return Promise.resolve(xmlApiData.mosObj)
-		})
-		onRequestAllMOSObjects = jest.fn(() => {
-			return Promise.resolve([
-				xmlApiData.mosObj,
-				xmlApiData.mosObj2
-			])
-		})
-		mosDevice.onRequestMOSObject((objId: string): Promise<IMOSObject | null> => {
-			return onRequestMOSObject(objId)
-		})
-		mosDevice.onRequestAllMOSObjects((): Promise<Array<IMOSObject>> => {
-			return onRequestAllMOSObjects()
-		})
-		let b = doBeforeAll()
-		socketMockLower = b.socketMockLower
-		socketMockUpper = b.socketMockUpper
-		socketMockQuery = b.socketMockQuery
-		serverSocketMockLower = b.serverSocketMockLower
-		serverSocketMockUpper = b.serverSocketMockUpper
-		serverSocketMockQuery = b.serverSocketMockQuery
-	})
-	afterAll(async () => {
-		await mosDevice.dispose()
-		await mosConnection.dispose()
-	})
-	beforeEach(() => {
-		onRequestMOSObject.mockClear()
-		onRequestAllMOSObjects.mockClear()
-
-		serverSocketMockLower.mockClear()
-		serverSocketMockUpper.mockClear()
-		if (serverSocketMockQuery) serverSocketMockQuery.mockClear()
-		socketMockLower.mockClear()
-		socketMockUpper.mockClear()
-		if (socketMockQuery) socketMockQuery.mockClear()
-	})
-	test('init', async () => {
-		expect(mosDevice).toBeTruthy()
-		expect(socketMockLower).toBeTruthy()
-		expect(socketMockUpper).toBeTruthy()
-		expect(serverSocketMockLower).toBeTruthy()
-	})
-	test('heartbeat from other party', async () => {
-
-		expect(serverSocketMockLower).toBeTruthy()
-
-		serverSocketMockLower.setReplyToHeartBeat(false)
-
-		let serverReply: jest.Mock<any, any> = jest.fn(() => false)
-		serverSocketMockLower.mockAddReply(serverReply)
-		// Fake incoming message on socket:
-		let sendMessageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.heartbeat)
-		await delay(10) // to allow for async timers & events to triggered
-
-		expect(serverReply).toHaveBeenCalledTimes(1)
-
-		let msg = serverSocketMockLower.decode(serverReply.mock.calls[0][0])
-
-		expect(msg).toMatch(/<heartbeat/)
-		expect(msg).toMatch('<messageID>' + sendMessageId)
-	})
-
-	test('unknown party connects', async () => {
-
-		// let unknownServerSocketMockLower = serverMockLower.mockNewConnection()
-
-		expect(serverSocketMockLower).toBeTruthy()
-		serverSocketMockLower.setReplyToHeartBeat(false)
-		let serverReply: jest.Mock<any, any> = jest.fn(() => false)
-		serverSocketMockLower.mockAddReply(serverReply)
-
-		// Fake incoming message on socket:
-		let sendMessageId = await fakeIncomingMessage(
-			serverSocketMockLower,
-			xmlData.heartbeat,
-			'ourUnknownMosId'
-		)
-		await delay(10) // to allow for async timers & events to triggered
-
-		expect(serverReply).toHaveBeenCalledTimes(1)
-		let msg = serverSocketMockLower.decode(serverReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosAck>/)
-		expect(msg).toMatch('<messageID>' + sendMessageId)
-		expect(msg).toMatch('<status>NACK')
-
-		serverReply.mockClear()
-		serverSocketMockLower.mockAddReply(serverReply)
-
-		// Fake incoming message on socket:
-		sendMessageId = await fakeIncomingMessage(
-			serverSocketMockLower,
-			xmlData.heartbeat,
-			undefined,
-			'theirUnknownMosId'
-		)
-		await delay(10) // to allow for async timers & events to triggered
-
-		expect(serverReply).toHaveBeenCalledTimes(1)
-		msg = serverSocketMockLower.decode(serverReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosAck>/)
-		expect(msg).toMatch('<messageID>' + sendMessageId)
-		expect(msg).toMatch('<status>NACK')
-
-	})
-	// TODO: reqMachInfo
-	// TODO: listMachInfo
-})
-
-describe('Profile 1', () => {
-	let mosDevice: MosDevice
-	let mosConnection: MosConnection
-
-	let socketMockLower: SocketMock
-	let socketMockUpper: SocketMock
-	let socketMockQuery: SocketMock
-
-	let serverSocketMockLower: SocketMock
-	let serverSocketMockUpper: SocketMock
-	let serverSocketMockQuery: SocketMock
-
-	let onRequestMachineInfo: jest.Mock<any, any>
-	let onRequestMOSObject: jest.Mock<any, any>
-	let onRequestAllMOSObjects: jest.Mock<any, any>
-
-	beforeAll(async () => {
-
-		mosConnection = await getMosConnection({
-			'0': true,
-			'1': true
-		}, true)
-		mosDevice = await getMosDevice(mosConnection)
-
-		// Profile 0:
-		onRequestMachineInfo = jest.fn(() => {
-			return Promise.resolve(xmlApiData.machineInfo)
-		})
-		mosDevice.onRequestMachineInfo((): Promise<IMOSListMachInfo> => {
-			return onRequestMachineInfo()
-		})
-		// Profile 1:
-		onRequestMOSObject = jest.fn(() => {
-			return Promise.resolve(xmlApiData.mosObj)
-		})
-		onRequestAllMOSObjects = jest.fn(() => {
-			return Promise.resolve([
-				xmlApiData.mosObj,
-				xmlApiData.mosObj2
-			])
-		})
-		mosDevice.onRequestMOSObject((objId: string): Promise<IMOSObject | null> => {
-			return onRequestMOSObject(objId)
-		})
-		mosDevice.onRequestAllMOSObjects((): Promise<Array<IMOSObject>> => {
-			return onRequestAllMOSObjects()
-		})
-		let b = doBeforeAll()
-		socketMockLower = b.socketMockLower
-		socketMockUpper = b.socketMockUpper
-		socketMockQuery = b.socketMockQuery
-		serverSocketMockLower = b.serverSocketMockLower
-		serverSocketMockUpper = b.serverSocketMockUpper
-		serverSocketMockQuery = b.serverSocketMockQuery
-	})
-	afterAll(async () => {
-		await mosDevice.dispose()
-		await mosConnection.dispose()
-	})
-	beforeEach(() => {
-		// SocketMock.mockClear()
-		onRequestMOSObject.mockClear()
-		onRequestAllMOSObjects.mockClear()
-
-		serverSocketMockLower.mockClear()
-		serverSocketMockUpper.mockClear()
-		if (serverSocketMockQuery) serverSocketMockQuery.mockClear()
-		socketMockLower.mockClear()
-		socketMockUpper.mockClear()
-		if (socketMockQuery) socketMockQuery.mockClear()
-	})
-	test('init', async () => {
-		expect(mosDevice).toBeTruthy()
-		expect(socketMockLower).toBeTruthy()
-		expect(socketMockUpper).toBeTruthy()
-		mosDevice.checkProfileValidness()
-	})
-
-	test('onRequestMOSObject', async () => {
-		// Fake incoming message on socket:
-		await fakeIncomingMessage(serverSocketMockLower, xmlData.reqObj)
-		expect(onRequestMOSObject).toHaveBeenCalledTimes(1)
-		expect(onRequestMOSObject.mock.calls[0][0]).toEqual(xmlApiData.mosObj.ID!.toString())
-		expect(onRequestMOSObject.mock.calls).toMatchSnapshot()
-		expect(onRequestAllMOSObjects).toHaveBeenCalledTimes(0)
-
-		// Check reply to socket server:
-		await serverSocketMockLower.mockWaitForSentMessages()
-		expect(serverSocketMockLower.mockSentMessage).toHaveBeenCalledTimes(1)
-		// @ts-ignore mock
-		let reply = decode(serverSocketMockLower.mockSentMessage['mock'].calls![0][0])
-		let parsedReply: any = xml2js(reply, { compact: true, nativeType: true, trim: true })
-
-		expect(parsedReply.mos.mosObj.objID._text + '').toEqual(xmlApiData.mosObj.ID!.toString())
-		expect(parsedReply.mos.mosObj.objSlug._text + '').toEqual(xmlApiData.mosObj.Slug.toString())
-		expect(parsedReply).toMatchSnapshot()
-	})
-	test('onRequestAllMOSObjects', async () => {
-
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			let repl = getXMLReply(messageID, '<mosAck></mosAck>')
-
-			return encode(repl)
-		})
-		socketMockLower.mockAddReply(mockReply)
-
-		// Fake incoming message on socket:
-		await fakeIncomingMessage(serverSocketMockLower, xmlData.mosReqAll)
-		expect(onRequestAllMOSObjects).toHaveBeenCalledTimes(1)
-		expect(onRequestAllMOSObjects.mock.calls).toMatchSnapshot()
-		expect(onRequestMOSObject).toHaveBeenCalledTimes(0)
-
-		// Check reply to socket server:
-		await serverSocketMockLower.mockWaitForSentMessages()
-		expect(serverSocketMockLower.mockSentMessage).toHaveBeenCalledTimes(1)
-		// @ts-ignore mock
-		let reply = decode(serverSocketMockLower.mockSentMessage.mock.calls[0][0])
-		let parsedReply: any = xml2js(reply, { compact: true, nativeType: true, trim: true })
-		expect(parsedReply.mos.mosAck).toBeTruthy()
-		expect(parsedReply).toMatchSnapshot()
-
-		// @ts-ignore mock
-		serverSocketMockLower.mockSentMessage.mockClear()
-		await delay(50)
-		await serverSocketMockLower.mockWaitForSentMessages()
-
-		expect(mockReply).toHaveBeenCalledTimes(1)
-
-		const sentData = Utils.xml2js(decode(mockReply.mock.calls[0][0])) as any
-
-		expect(sentData.mos.mosListAll.mosObj).toHaveLength(2)
-		expect(sentData.mos.mosListAll.mosObj[0].objID + '').toEqual(xmlApiData.mosObj.ID!.toString())
-		expect(sentData.mos.mosListAll.mosObj[0].objSlug + '').toEqual(xmlApiData.mosObj.Slug.toString())
-		expect(sentData.mos.mosListAll.mosObj[1].objID + '').toEqual(xmlApiData.mosObj2.ID!.toString())
-		expect(sentData.mos.mosListAll.mosObj[1].objSlug + '').toEqual(xmlApiData.mosObj2.Slug.toString())
-		sentData.mos.messageID = 99999 // not important
-		expect(sentData).toMatchSnapshot()
-	})
-	test('getMOSObject', async () => {
-
-		// Prepare mock server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			let repl = getXMLReply(messageID, xmlData.mosObj)
-
-			return encode(repl)
-		})
-		socketMockLower.mockAddReply(mockReply)
-		let returnedObj: IMOSObject = await mosDevice.sendRequestMOSObject(xmlApiData.mosObj.ID!)
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqObj>/)
-		checkMessageSnapshot(msg)
-
-		expect(returnedObj).toMatchObject(xmlApiData.roList)
-		expect(returnedObj).toMatchSnapshot()
-	})
-	test('getAllMOSObjects', async () => {
-
-		expect(socketMockLower).toBeTruthy()
-		// Prepare mock server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			let repl = getXMLReply(messageID, xmlData.mosListAll)
-			return encode(repl)
-		})
-		socketMockLower.mockAddReply(mockReply)
-		let returnedObjs: Array<IMOSObject> = await mosDevice.sendRequestAllMOSObjects()
-
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqAll>/)
-		checkMessageSnapshot(msg)
-
-		expect(returnedObjs).toMatchObject(xmlApiData.mosListAll)
-		expect(returnedObjs).toMatchSnapshot()
-	})
-})
-
-describe('Profile 2', () => {
-	let mosDevice: MosDevice
-	let mosConnection: MosConnection
-	let socketMockLower: SocketMock
-	let socketMockUpper: SocketMock
-
-	let serverSocketMockLower: SocketMock
-	let serverSocketMockUpper: SocketMock
-	let serverSocketMockQuery: SocketMock
-
-	let onRequestMachineInfo: jest.Mock<any, any>
-	let onRequestMOSObject: jest.Mock<any, any>
-	let onRequestAllMOSObjects: jest.Mock<any, any>
-	let onCreateRunningOrder: jest.Mock<any, any>
-	let onReplaceRunningOrder: jest.Mock<any, any>
-	let onDeleteRunningOrder: jest.Mock<any, any>
-	let onRequestRunningOrder: jest.Mock<any, any>
-	let onMetadataReplace: jest.Mock<any, any>
-	let onRunningOrderStatus: jest.Mock<any, any>
-	let onStoryStatus: jest.Mock<any, any>
-	let onItemStatus: jest.Mock<any, any>
-	let onReadyToAir: jest.Mock<any, any>
-	let onROInsertStories: jest.Mock<any, any>
-	let onROInsertItems: jest.Mock<any, any>
-	let onROReplaceStories: jest.Mock<any, any>
-	let onROReplaceItems: jest.Mock<any, any>
-	let onROMoveStories: jest.Mock<any, any>
-	let onROMoveItems: jest.Mock<any, any>
-	let onRODeleteStories: jest.Mock<any, any>
-	let onRODeleteItems: jest.Mock<any, any>
-	let onROSwapStories: jest.Mock<any, any>
-	let onROSwapItems: jest.Mock<any, any>
-
-	beforeAll(async () => {
-		SocketMock.mockClear()
-		ServerMock.mockClear()
-
-		mosConnection = await getMosConnection({
-			'0': true,
-			'1': true,
-			'2': true
-		}, true)
-		mosDevice = await getMosDevice(mosConnection)
-
-		// Profile 0:
-		onRequestMachineInfo = jest.fn(() => {
-			return Promise.resolve(xmlApiData.machineInfo)
-		})
-		mosDevice.onRequestMachineInfo((): Promise<IMOSListMachInfo> => {
-			return onRequestMachineInfo()
-		})
-		// Profile 1:
-		onRequestMOSObject = jest.fn()
-		onRequestAllMOSObjects = jest.fn()
-		mosDevice.onRequestMOSObject((objId: string): Promise<IMOSObject | null> => {
-			return onRequestMOSObject(objId)
-		})
-		mosDevice.onRequestAllMOSObjects((): Promise<Array<IMOSObject>> => {
-			return onRequestAllMOSObjects()
-		})
-
-		// Profile 2:
-		const roAckReply = () => {
-			let ack: IMOSROAck = {
-				ID: new MosString128('runningOrderId'),
-				Status: new MosString128('OK'),
-				Stories: []
-			}
-			return Promise.resolve(ack)
-		}
-		onCreateRunningOrder = jest.fn(roAckReply)
-		onReplaceRunningOrder = jest.fn(roAckReply)
-		onDeleteRunningOrder = jest.fn(roAckReply)
-		onRequestRunningOrder = jest.fn(() => {
-			return Promise.resolve(xmlApiData.roCreate)
-		})
-		onMetadataReplace = jest.fn(roAckReply)
-		onRunningOrderStatus = jest.fn(roAckReply)
-		onStoryStatus = jest.fn(roAckReply)
-		onItemStatus = jest.fn(roAckReply)
-		onReadyToAir = jest.fn(roAckReply)
-		onROInsertStories = jest.fn(roAckReply)
-		onROInsertItems = jest.fn(roAckReply)
-		onROReplaceStories = jest.fn(roAckReply)
-		onROReplaceItems = jest.fn(roAckReply)
-		onROMoveStories = jest.fn(roAckReply)
-		onROMoveItems = jest.fn(roAckReply)
-		onRODeleteStories = jest.fn(roAckReply)
-		onRODeleteItems = jest.fn(roAckReply)
-		onROSwapStories = jest.fn(roAckReply)
-		onROSwapItems = jest.fn(roAckReply)
-
-		mosDevice.onCreateRunningOrder((ro: IMOSRunningOrder): Promise<IMOSROAck> => {
-			return onCreateRunningOrder(ro)
-		})
-		mosDevice.onReplaceRunningOrder((ro: IMOSRunningOrder): Promise<IMOSROAck> => {
-			return onReplaceRunningOrder(ro)
-		})
-		mosDevice.onDeleteRunningOrder((runningOrderId: MosString128): Promise<IMOSROAck> => {
-			return onDeleteRunningOrder(runningOrderId)
-		})
-		mosDevice.onRequestRunningOrder((runningOrderId: MosString128): Promise<IMOSRunningOrder | null> => {
-			return onRequestRunningOrder(runningOrderId)
-		})
-		mosDevice.onMetadataReplace((metadata: IMOSRunningOrderBase): Promise<IMOSROAck> => {
-			return onMetadataReplace(metadata)
-		})
-		mosDevice.onRunningOrderStatus((status: IMOSRunningOrderStatus): Promise<IMOSROAck> => {
-			return onRunningOrderStatus(status)
-		})
-		mosDevice.onStoryStatus((status: IMOSStoryStatus): Promise<IMOSROAck> => {
-			return onStoryStatus(status)
-		})
-		mosDevice.onItemStatus((status: IMOSItemStatus): Promise<IMOSROAck> => {
-			return onItemStatus(status)
-		})
-		mosDevice.onReadyToAir((Action: IMOSROReadyToAir): Promise<IMOSROAck> => {
-			return onReadyToAir(Action)
-		})
-		mosDevice.onROInsertStories((Action: IMOSStoryAction, Stories: Array<IMOSROStory>): Promise<IMOSROAck> => {
-			return onROInsertStories(Action, Stories)
-		})
-		mosDevice.onROInsertItems((Action: IMOSItemAction, Items: Array<IMOSItem>): Promise<IMOSROAck> => {
-			return onROInsertItems(Action, Items)
-		})
-		mosDevice.onROReplaceStories((Action: IMOSStoryAction, Stories: Array<IMOSROStory>): Promise<IMOSROAck> => {
-			return onROReplaceStories(Action, Stories)
-		})
-		mosDevice.onROReplaceItems((Action: IMOSItemAction, Items: Array<IMOSItem>): Promise<IMOSROAck> => {
-			return onROReplaceItems(Action, Items)
-		})
-		mosDevice.onROMoveStories((Action: IMOSStoryAction, Stories: Array<MosString128>): Promise<IMOSROAck> => {
-			return onROMoveStories(Action, Stories)
-		})
-		mosDevice.onROMoveItems((Action: IMOSItemAction, Items: Array<MosString128>): Promise<IMOSROAck> => {
-			return onROMoveItems(Action, Items)
-		})
-		mosDevice.onRODeleteStories((Action: IMOSROAction, Stories: Array<MosString128>): Promise<IMOSROAck> => {
-			return onRODeleteStories(Action, Stories)
-		})
-		mosDevice.onRODeleteItems((Action: IMOSStoryAction, Items: Array<MosString128>): Promise<IMOSROAck> => {
-			return onRODeleteItems(Action, Items)
-		})
-		mosDevice.onROSwapStories((Action: IMOSROAction, StoryID0: MosString128, StoryID1: MosString128): Promise<IMOSROAck> => {
-			return onROSwapStories(Action, StoryID0, StoryID1)
-		})
-		mosDevice.onROSwapItems((Action: IMOSStoryAction, ItemID0: MosString128, ItemID1: MosString128): Promise<IMOSROAck> => {
-			return onROSwapItems(Action, ItemID0, ItemID1)
-		})
-		let b = doBeforeAll()
-		socketMockLower = b.socketMockLower
-		socketMockUpper = b.socketMockUpper
-		serverSocketMockLower = b.serverSocketMockLower
-		serverSocketMockUpper = b.serverSocketMockUpper
-		serverSocketMockQuery = b.serverSocketMockQuery
-	})
-	beforeEach(() => {
-		onRequestMOSObject.mockClear()
-		onRequestAllMOSObjects.mockClear()
-
-		onCreateRunningOrder.mockClear()
-		onReplaceRunningOrder.mockClear()
-		onDeleteRunningOrder.mockClear()
-		onRequestRunningOrder.mockClear()
-		onMetadataReplace.mockClear()
-		onRunningOrderStatus.mockClear()
-		onStoryStatus.mockClear()
-		onItemStatus.mockClear()
-		onReadyToAir.mockClear()
-		onROInsertStories.mockClear()
-		onROInsertItems.mockClear()
-		onROReplaceStories.mockClear()
-		onROReplaceItems.mockClear()
-		onROMoveStories.mockClear()
-		onROMoveItems.mockClear()
-		onRODeleteStories.mockClear()
-		onRODeleteItems.mockClear()
-		onROSwapStories.mockClear()
-		onROSwapItems.mockClear()
-
-		serverSocketMockLower.mockClear()
-		serverSocketMockUpper.mockClear()
-		serverSocketMockQuery.mockClear()
-
-	})
-	test('init', async () => {
-		expect(mosDevice).toBeTruthy()
-		expect(socketMockLower).toBeTruthy()
-		expect(socketMockUpper).toBeTruthy()
-		mosDevice.checkProfileValidness()
-	})
-	test('onCreateRunningOrder', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roCreate)
-		expect(onCreateRunningOrder).toHaveBeenCalledTimes(1)
-		expect(onCreateRunningOrder.mock.calls[0][0]).toMatchObject(xmlApiData.roCreate)
-		expect(onCreateRunningOrder.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	// TODO: sendCreateRunningOrder
-	test('onReplaceRunningOrder', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roReplace)
-		expect(onReplaceRunningOrder).toHaveBeenCalledTimes(1)
-		expect(onReplaceRunningOrder.mock.calls[0][0]).toMatchObject(xmlApiData.roReplace)
-		expect(onReplaceRunningOrder.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	// TODO: sendReplaceRunningOrder
-	test('onDeleteRunningOrder', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roDelete)
-		expect(onDeleteRunningOrder).toHaveBeenCalledTimes(1)
-		expect(onDeleteRunningOrder.mock.calls[0][0]).toEqual(xmlApiData.roDelete)
-		expect(onDeleteRunningOrder.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	// TODO: sendDeleteRunningOrder
-	test('onRequestRunningOrder', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockUpper, xmlData.roReq)
-		expect(onRequestRunningOrder).toHaveBeenCalledTimes(1)
-
-		expect(onRequestRunningOrder.mock.calls[0][0]).toEqual(96857485)
-		expect(onRequestRunningOrder.mock.calls).toMatchSnapshot()
-		// Check reply to socket server:
-		await serverSocketMockUpper.mockWaitForSentMessages()
-		expect(serverSocketMockUpper.mockSentMessage).toHaveBeenCalledTimes(1)
-		await checkReplyToServer(serverSocketMockUpper, messageId, '<roList>')
-		// @ts-ignore mock
-		let reply = decode(serverSocketMockUpper.mockSentMessage.mock.calls[0][0])
-		let parsedReply: any = xml2js(reply, { compact: true, nativeType: true, trim: true })
-
-		expect(parsedReply.mos.roList.roID._text + '').toEqual(xmlApiData.roCreate.ID.toString())
-		expect(parsedReply.mos.roList.roSlug._text + '').toEqual(xmlApiData.roCreate.Slug.toString())
-		expect(parsedReply.mos.roList.story).toHaveLength(xmlApiData.roCreate.Stories.length)
-		expect(parsedReply.mos.roList.story[0].storyID._text + '').toEqual(xmlApiData.roCreate.Stories[0].ID.toString())
-		expect(parsedReply.mos.roList.story[0].item).toBeTruthy()
-		expect(parsedReply.mos.roList.story[0].item.itemID._text + '').toEqual(xmlApiData.roCreate.Stories[0].Items[0].ID.toString())
-		expect(parsedReply.mos.roList.story[0].item.objID._text + '').toEqual(xmlApiData.roCreate.Stories[0].Items[0].ObjectID.toString())
-
-		expect(parsedReply).toMatchSnapshot()
-
-	})
-	test('sendRequestRunningOrder', async () => {
-
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			let repl = getXMLReply(messageID, xmlData.roList)
-			return encode(repl)
-
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedObj = await mosDevice.sendRequestRunningOrder(xmlApiData.roList.ID!) as IMOSRunningOrder
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roReq>/)
-		checkMessageSnapshot(msg)
-		expect(returnedObj).toMatchObject(xmlApiData.roList2)
-		expect(returnedObj).toMatchSnapshot()
-	})
-	test('onMetadataReplace', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roMetadataReplace)
-		expect(onMetadataReplace).toHaveBeenCalledTimes(1)
-		expect(onMetadataReplace.mock.calls[0][0]).toEqual(xmlApiData.roMetadataReplace)
-		expect(onMetadataReplace.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	// TODO: sendMetadataReplace
-	test('onRunningOrderStatus', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementStat_ro)
-		expect(onRunningOrderStatus).toHaveBeenCalledTimes(1)
-		expect(onRunningOrderStatus.mock.calls[0][0]).toEqual(xmlApiData.roElementStat_ro)
-		expect(onRunningOrderStatus.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onStoryStatus', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementStat_story)
-		expect(onStoryStatus).toHaveBeenCalledTimes(1)
-		expect(onStoryStatus.mock.calls[0][0]).toEqual(xmlApiData.roElementStat_story)
-		expect(onStoryStatus.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	// TODO: sendStoryStatus
-	test('onItemStatus', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementStat_item)
-		expect(onItemStatus).toHaveBeenCalledTimes(1)
-		expect(onItemStatus.mock.calls[0][0]).toEqual(xmlApiData.roElementStat_item)
-		expect(onItemStatus.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('sendRunningOrderStatus', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			let repl = getXMLReply(messageID, xmlData.roAck)
-			return encode(repl)
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedAck: IMOSROAck = await mosDevice.sendRunningOrderStatus(xmlApiData.roElementStat_ro)
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roElementStat element="RO">/)
-		checkMessageSnapshot(msg)
-		expect(returnedAck).toBeTruthy()
-		expect(returnedAck.ID.toString()).toEqual('96857485')
-		checkAckSnapshot(returnedAck)
-	})
-	test('sendStoryStatus', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.roAck))
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedAck: IMOSROAck = await mosDevice.sendStoryStatus(xmlApiData.roElementStat_story)
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roElementStat element="STORY">/)
-		checkMessageSnapshot(msg)
-		expect(returnedAck).toBeTruthy()
-		expect(returnedAck.ID.toString()).toEqual('96857485')
-		checkAckSnapshot(returnedAck)
-	})
-	test('sendItemStatus', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.roAck))
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedAck: IMOSROAck = await mosDevice.sendItemStatus(xmlApiData.roElementStat_item)
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roElementStat element="ITEM">/)
-		checkMessageSnapshot(msg)
-		expect(returnedAck).toBeTruthy()
-		expect(returnedAck.ID.toString()).toEqual('96857485')
-		checkAckSnapshot(returnedAck)
-	})
-	test('onReadyToAir', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roReadyToAir)
-		expect(onReadyToAir).toHaveBeenCalledTimes(1)
-		expect(onReadyToAir.mock.calls[0][0]).toEqual(xmlApiData.roReadyToAir)
-		expect(onReadyToAir.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROInsertStories', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_insert_story)
-		expect(onROInsertStories).toHaveBeenCalledTimes(1)
-		expect(onROInsertStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_insert_story_Action)
-		expect(onROInsertStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_insert_story_Stories)
-		expect(onROInsertStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROInsertItems', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_insert_item)
-		expect(onROInsertItems).toHaveBeenCalledTimes(1)
-		expect(onROInsertItems.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_insert_item_Action)
-		expect(onROInsertItems.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_insert_item_Items)
-		expect(onROInsertItems.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROReplaceStories', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_replace_story)
-		expect(onROReplaceStories).toHaveBeenCalledTimes(1)
-		expect(onROReplaceStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_replace_story_Action)
-		expect(onROReplaceStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_replace_story_Stories)
-		expect(onROReplaceStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROReplaceItems', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_replace_item)
-		expect(onROReplaceItems).toHaveBeenCalledTimes(1)
-		expect(onROReplaceItems.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_replace_item_Action)
-		expect(onROReplaceItems.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_replace_item_Items)
-		expect(onROReplaceItems.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROMoveStory', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_move_story)
-		expect(onROMoveStories).toHaveBeenCalledTimes(1)
-		expect(onROMoveStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_move_story_Action)
-		expect(onROMoveStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_move_story_Stories)
-		expect(onROMoveStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROMoveStories', async () => {
-		// Fake incoming message on socket:
-		let messageId2 = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_move_stories)
-		expect(onROMoveStories).toHaveBeenCalledTimes(1)
-		expect(onROMoveStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_move_stories_Action)
-		expect(onROMoveStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_move_stories_Stories)
-		expect(onROMoveStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId2, '<roAck>')
-	})
-	test('onROMoveItems', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_move_items)
-		expect(onROMoveItems).toHaveBeenCalledTimes(1)
-		expect(onROMoveItems.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_move_items_Action)
-		expect(onROMoveItems.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_move_items_Items)
-		expect(onROMoveItems.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onRODeleteStories', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_delete_story)
-		expect(onRODeleteStories).toHaveBeenCalledTimes(1)
-		expect(onRODeleteStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_delete_story_Action)
-		expect(onRODeleteStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_delete_story_Stories)
-		expect(onRODeleteStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onRODeleteItems', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_delete_items)
-		expect(onRODeleteItems).toHaveBeenCalledTimes(1)
-		expect(onRODeleteItems.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_delete_items_Action)
-		expect(onRODeleteItems.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_delete_items_Items)
-		expect(onRODeleteItems.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROSwapStories', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_swap_stories)
-		expect(onROSwapStories).toHaveBeenCalledTimes(1)
-		expect(onROSwapStories.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_swap_stories_Action)
-		expect(onROSwapStories.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_swap_stories_StoryId0)
-		expect(onROSwapStories.mock.calls[0][2]).toEqual(xmlApiData.roElementAction_swap_stories_StoryId1)
-		expect(onROSwapStories.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-	test('onROSwapItems', async () => {
-		// Fake incoming message on socket:
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roElementAction_swap_items)
-		expect(onROSwapItems).toHaveBeenCalledTimes(1)
-		expect(onROSwapItems.mock.calls[0][0]).toEqual(xmlApiData.roElementAction_swap_items_Action)
-		expect(onROSwapItems.mock.calls[0][1]).toEqual(xmlApiData.roElementAction_swap_items_ItemId0)
-		expect(onROSwapItems.mock.calls[0][2]).toEqual(xmlApiData.roElementAction_swap_items_ItemId1)
-		expect(onROSwapItems.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-	})
-})
-
-describe('Profile 3', () => {
-	let mosDevice: MosDevice
-	let mosConnection: MosConnection
-	let socketMockLower: SocketMock
-	let socketMockUpper: SocketMock
-	let socketMockQuery: SocketMock
-
-	let serverSocketMockLower: SocketMock
-	let serverSocketMockUpper: SocketMock
-	let serverSocketMockQuery: SocketMock
-
-	let onMosObjCreate: jest.Mock<any, any>
-	let onMosItemReplace: jest.Mock<any, any>
-	let onMosReqSearchableSchema: jest.Mock<any, any>
-	let onMosReqObjectList: jest.Mock<any, any>
-	let onRequestObjectActionNew: jest.Mock<any, any>
-	let onRequestObjectActionUpdate: jest.Mock<any, any>
-	let onRequestObjectActionDelete: jest.Mock<any, any>
-
-	let roAckReply = () => {
-		let ack: IMOSROAck = {
-			ID: new MosString128('runningOrderId'),
-			Status: new MosString128('OK'),
-			Stories: []
-		}
-		return Promise.resolve(ack)
-	}
-
-	let mosAckReply = () => {
-		let ack: IMOSAck = {
-			ID: new MosString128('runningOrderId'),
-			Revision: 1,
-			Status: IMOSAckStatus.ACK,
-			Description: new MosString128('')
-		}
-		return Promise.resolve(ack)
-	}
-
-	beforeAll(async () => {
-		SocketMock.mockClear()
-		ServerMock.mockClear()
-
-		mosConnection = await getMosConnection({
-			'0': true,
-			'1': true,
-			'2': true,
-			'3': true
-		}, false)
-		mosDevice = await getMosDevice(mosConnection)
-
-		// Profile 3:
-
-		onMosObjCreate = jest.fn(mosAckReply)
-		onMosItemReplace = jest.fn(roAckReply)
-		onMosReqSearchableSchema = jest.fn((username: string) => {
-			let response: IMOSListSearchableSchema = {
-				username,
-				mosSchema: 'http://example.com/mosSearchableSchema'
-			}
-			return Promise.resolve(response)
-		})
-		onMosReqObjectList = jest.fn((options: IMOSRequestObjectList) => {
-			let response: IMOSObjectList = {
-				username: options.username,
-				queryID: 'A392938329kdakd2039300d0s9l3l9d0bzAQ',
-				listReturnStart: 0,
-				listReturnEnd: 0,
-				listReturnTotal: 0
-			}
-			return Promise.resolve(response)
-		})
-		onRequestObjectActionNew = jest.fn(mosAckReply)
-		onRequestObjectActionUpdate = jest.fn(mosAckReply)
-		onRequestObjectActionDelete = jest.fn(mosAckReply)
-
-		mosDevice.onObjectCreate((obj: IMOSObject) => {
-			return onMosObjCreate(obj)
-		})
-		mosDevice.onItemReplace((roID, storyID, item) => {
-			return onMosItemReplace(roID, storyID, item)
-		})
-		mosDevice.onRequestSearchableSchema((username) => {
-			return onMosReqSearchableSchema(username)
-		})
-		mosDevice.onRequestObjectList((options: IMOSRequestObjectList) => {
-			return onMosReqObjectList(options)
-		})
-		mosDevice.onRequestObjectActionNew((obj) => {
-			return onRequestObjectActionNew(obj)
-		})
-		mosDevice.onRequestObjectActionUpdate((objId, obj) => {
-			return onRequestObjectActionUpdate(objId, obj)
-		})
-		mosDevice.onRequestObjectActionDelete((objId) => {
-			return onRequestObjectActionDelete(objId)
-		})
-		let b = doBeforeAll()
-		socketMockLower = b.socketMockLower
-		socketMockUpper = b.socketMockUpper
-		socketMockQuery = b.socketMockQuery
-		serverSocketMockLower = b.serverSocketMockLower
-		serverSocketMockUpper = b.serverSocketMockUpper
-		serverSocketMockQuery = b.serverSocketMockQuery
-	})
-	beforeEach(() => {
-		onMosObjCreate.mockClear()
-		onMosItemReplace.mockClear()
-		onMosReqSearchableSchema.mockClear()
-		onMosReqObjectList.mockClear()
-		onRequestObjectActionNew.mockClear()
-		onRequestObjectActionUpdate.mockClear()
-		onRequestObjectActionDelete.mockClear()
-
-		serverSocketMockLower.mockClear()
-		serverSocketMockUpper.mockClear()
-		serverSocketMockQuery.mockClear()
-	})
-	test('init', async () => {
-		expect(mosDevice).toBeTruthy()
-		expect(socketMockLower).toBeTruthy()
-		expect(socketMockUpper).toBeTruthy()
-		expect(socketMockQuery).toBeTruthy()
-		mosDevice.checkProfileValidness()
-	})
-	test('onMosObjCreate', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.mosObjCreate)
-		expect(onMosObjCreate).toHaveBeenCalledTimes(1)
-		expect(onMosObjCreate.mock.calls[0][0]).toMatchObject(xmlApiData.mosObjCreate)
-		expect(onMosObjCreate.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockLower, messageId, '<mosAck>')
-	})
-	test('mosObjCreate', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosAck))
-		})
-		socketMockLower.mockAddReply(mockReply)
-		let returnedAck = await mosDevice.sendObjectCreate({
-			ID: new MosString128('abc'),
-			Slug: new MosString128('my cool Object'),
-			Type: IMOSObjectType.VIDEO,
-			TimeBase: 25,
-			Duration: 250
-		})
-		await socketMockLower.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosObjCreate>/)
-		checkMessageSnapshot(msg)
-		expect(returnedAck).toBeTruthy()
-		checkAckSnapshot(returnedAck)
-	})
-	test('onMosItemReplace', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockUpper, xmlData.mosItemReplace)
-		expect(onMosItemReplace).toHaveBeenCalledTimes(1)
-		expect(onMosItemReplace.mock.calls[0][2]).toMatchObject(xmlApiData.mosItemReplace)
-		expect(onMosItemReplace.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockUpper, messageId, '<roAck>')
-	})
-	test('mosItemReplace', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.roAck))
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedAck: IMOSROAck = await mosDevice.sendItemReplace({
-			roID: new MosString128('roX'),
-			storyID: new MosString128('storyY'),
-			item: {
-				ID: new MosString128('abc'),
-				Slug: new MosString128('my cool Object'),
-				// Type: IMOSObjectType.VIDEO,
-				TimeBase: 25,
-				Duration: 250,
-				ObjectID: new MosString128('Object0'),
-				MOSID: 'our.mos.id'
-			}
-		})
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosItemReplace>/)
-		expect(msg).toMatch(/<roID>roX<\/roID>/)
-		expect(msg).toMatch(/<storyID>storyY<\/storyID>/)
-		checkMessageSnapshot(msg)
-		expect(returnedAck).toBeTruthy()
-		checkAckSnapshot(returnedAck)
-	})
-	test('onMosReqSearchableSchema', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockQuery, xmlData.mosReqSearchableSchema)
-		expect(onMosReqSearchableSchema).toHaveBeenCalledTimes(1)
-		expect(onMosReqSearchableSchema.mock.calls[0][0]).toMatch(xmlApiData.mosReqSearchableSchema)
-		expect(onMosReqSearchableSchema.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockQuery, messageId, '<mosListSearchableSchema username="jbob"')
-	})
-	test('mosReqSearchableSchema', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosListSearchableSchema))
-		})
-		socketMockQuery.mockAddReply(mockReply)
-		let returnedSchema: IMOSListSearchableSchema = await mosDevice.sendRequestSearchableSchema('myUsername')
-
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqSearchableSchema username="myUsername"/)
-		checkMessageSnapshot(msg)
-
-		expect(returnedSchema.username).toEqual('myUsername')
-		expect(returnedSchema.mosSchema).toEqual('http://MOSA4.com/mos/supported_schemas/MOSAXML2.08')
-		expect(returnedSchema).toMatchSnapshot()
-	})
-	test('onMosReqObjectList', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockQuery, xmlData.mosReqObjList)
-		expect(onMosReqObjectList).toHaveBeenCalledTimes(1)
-		expect(onMosReqObjectList.mock.calls[0][0]).toMatchObject(xmlApiData.mosReqObjList)
-		expect(onMosReqObjectList.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockQuery, messageId, '<mosObjList')
-	})
-	test('mosRequestObjectList', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosObjList))
-		})
-		socketMockQuery.mockAddReply(mockReply)
-		let returnedObjList: IMOSObjectList = await mosDevice.sendRequestObjectList({
-			username: 'jbob',
-			queryID: new MosString128('A392938329kdakd2039300d0s9l3l9d0bzAQ'),
-			listReturnStart: 1,
-			listReturnEnd: 20,
-			generalSearch: new MosString128('boats'),
-			mosSchema: '',
-			searchGroups: []
-		})
-
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqObjList username="jbob">/)
-		checkMessageSnapshot(msg)
-
-		expect(returnedObjList.username).toEqual('jbob')
-		expect(returnedObjList.list).toHaveLength(3)
-		expect(returnedObjList.list![0]).toMatchObject({
-			ID: new MosString128('M000121'),
-			Slug: new MosString128('Hotel Fire'),
-			Group: 'Show 7',
-			Type: IMOSObjectType.VIDEO,
-			TimeBase: 59.94,
-			Revision: 1,
-			Duration: 1800,
-			Status: IMOSObjectStatus.NEW,
-			AirStatus: IMOSObjectAirStatus.READY
-		})
-		expect(returnedObjList).toMatchSnapshot()
-	})
-	test('onMosReqObjectActionNew', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockQuery, xmlData.mosReqObjActionNew)
-		expect(onRequestObjectActionNew).toHaveBeenCalledTimes(1)
-		expect(onRequestObjectActionNew.mock.calls[0][0]).toMatchObject(xmlApiData.mosObjReqObjActionNew)
-		expect(onRequestObjectActionNew.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockQuery, messageId, '<mosAck>')
-	})
-	test('onMosReqObjectActionUpdate', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockQuery, xmlData.mosReqObjActionUpdate)
-		expect(onRequestObjectActionUpdate).toHaveBeenCalledTimes(1)
-		expect(onRequestObjectActionUpdate.mock.calls[0][0]).toBe(xmlApiData.mosObjReqObjActionUpdateObjId)
-		expect(onRequestObjectActionUpdate.mock.calls[0][1]).toMatchObject(xmlApiData.mosObjReqObjActionUpdate)
-		expect(onRequestObjectActionUpdate.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockQuery, messageId, '<mosAck>')
-	})
-	test('onMosReqObjectActionDelete ', async () => {
-		let messageId = await fakeIncomingMessage(serverSocketMockQuery, xmlData.mosReqObjActionDelete)
-		expect(onRequestObjectActionDelete).toHaveBeenCalledTimes(1)
-		expect(onRequestObjectActionDelete.mock.calls[0][0]).toBe(xmlApiData.mosObjReqObjActionDeleteObjId)
-		expect(onRequestObjectActionDelete.mock.calls).toMatchSnapshot()
-		await checkReplyToServer(serverSocketMockQuery, messageId, '<mosAck>')
-	})
-	test('sendMosReqObjectActionNew', async () => {
-		// Prepare server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosAck))
-		})
-		socketMockLower.mockAddReply(mockReply)
-		await mosDevice.sendRequestObjectActionNew({
-			Slug: new MosString128('abc123'),
-			Type: IMOSObjectType.VIDEO,
-			TimeBase: 25,
-			Duration: 500
-		})
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqObjAction operation="NEW">/)
-		checkMessageSnapshot(msg)
-	})
-	test('sendMosReqObjectActionUpdateUpdate', async () => {
-		// Prepare server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosAck))
-		})
-		socketMockLower.mockAddReply(mockReply)
-		await mosDevice.sendRequestObjectActionUpdate(new MosString128('OBJID1234'),{
-			Slug: new MosString128('abc123'),
-			Type: IMOSObjectType.VIDEO,
-			TimeBase: 25,
-			Duration: 500
-		})
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqObjAction operation="UPDATE" objID="OBJID1234">/)
-		checkMessageSnapshot(msg)
-	})
-	test('sendMosReqObjectActionUpdateDelete', async () => {
-		// Prepare server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.mosAck))
-		})
-		socketMockLower.mockAddReply(mockReply)
-		await mosDevice.sendRequestObjectActionDelete(new MosString128('OBJID1234'))
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<mosReqObjAction operation="DELETE" objID="OBJID1234"\/>/)
-		checkMessageSnapshot(msg)
-	})
-	test('sendRunningOrderStory', async () => {
-		// Prepare server response:
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.roAck))
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		await mosDevice.sendRunningOrderStory(xmlApiData.sendRunningOrderStory)
-		await socketMockQuery.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roStorySend>/)
-		expect(msg).toMatch(/<roID>96857485<\/roID>/)
-		expect(msg).toMatch(/<storyID>5983A501:0049B924:8390EF1F<\/storyID>/)
-		checkMessageSnapshot(msg)
-	})
-
-})
-
-describe('Profile 4', () => {
-	let mosDevice: MosDevice
-	let mosConnection: MosConnection
-	let socketMockLower: SocketMock
-	let socketMockUpper: SocketMock
-
-	let serverSocketMockLower: SocketMock
-	let serverSocketMockUpper: SocketMock
-	let serverSocketMockQuery: SocketMock
-
-	let onROStory: jest.Mock<any, any>
-
-	beforeAll(async () => {
-		SocketMock.mockClear()
-		ServerMock.mockClear()
-
-		mosConnection = await getMosConnection({
-			'0': true,
-			'1': true,
-			'2': true,
-			'4': true
-		}, false)
-		mosDevice = await getMosDevice(mosConnection)
-
-		let roAckReply = () => {
-			let ack: IMOSROAck = {
-				ID: new MosString128('runningOrderId'),
-				Status: new MosString128('OK'),
-				Stories: []
-			}
-			return Promise.resolve(ack)
-		}
-		// Profile 2:
-		onROStory = jest.fn(roAckReply)
-
-		mosDevice.onRunningOrderStory((story: IMOSROFullStory): Promise<IMOSROAck> => {
-			return onROStory(story)
-		})
-		let b = doBeforeAll()
-		socketMockLower = b.socketMockLower
-		socketMockUpper = b.socketMockUpper
-		serverSocketMockLower = b.serverSocketMockLower
-		serverSocketMockUpper = b.serverSocketMockUpper
-		serverSocketMockQuery = b.serverSocketMockQuery
-	})
-	beforeEach(() => {
-
-		onROStory.mockClear()
-
-		serverSocketMockLower.mockClear()
-		serverSocketMockUpper.mockClear()
-		serverSocketMockQuery.mockClear()
-
-	})
-	test('init', async () => {
-		expect(mosDevice).toBeTruthy()
-		expect(socketMockLower).toBeTruthy()
-		expect(socketMockUpper).toBeTruthy()
-		mosDevice.checkProfileValidness()
-	})
-	test('onROStory', async () => {
-		// Fake incoming message on socket:
-
-		let messageId = await fakeIncomingMessage(serverSocketMockLower, xmlData.roStorySend)
-		expect(onROStory).toHaveBeenCalledTimes(1)
-
-		let o = Object.assign({}, xmlApiData.roStorySend)
-		// @ts-ignore optional property
-		delete o.Body
-		expect(onROStory.mock.calls[0][0]).toMatchObject(o)
-		expect(onROStory.mock.calls).toMatchSnapshot()
-		xmlApiData.roStorySend.Body.forEach((testItem, key) => {
-			let item: any
-			try {
-				item = onROStory.mock.calls[0][0].Body[key]
-				if (!testItem.Content) delete testItem.Content
-
-				expect(item).toMatchObject(testItem)
-			} catch (e) {
-				console.error(key)
-				console.error('item', item)
-				console.error('testItem', testItem)
-				throw e
-			}
-		})
-		await checkReplyToServer(serverSocketMockLower, messageId, '<roAck>')
-
-	})
-	test('getAllRunningOrders', async () => {
-		// Prepare server response
-		let mockReply = jest.fn((data) => {
-			let str = decode(data)
-			let messageID = str.match(/<messageID>([^<]+)<\/messageID>/)![1]
-			return encode(getXMLReply(messageID, xmlData.roListAll))
-		})
-		socketMockUpper.mockAddReply(mockReply)
-		let returnedListAll = await mosDevice.sendRequestAllRunningOrders()
-		await socketMockUpper.mockWaitForSentMessages()
-		expect(mockReply).toHaveBeenCalledTimes(1)
-		let msg = decode(mockReply.mock.calls[0][0])
-		expect(msg).toMatch(/<roReqAll/)
-		checkMessageSnapshot(msg)
-
-		expect(returnedListAll).toHaveLength(2)
-		expect(returnedListAll[0]).toMatchObject(xmlApiData.roListAll[0])
-		expect(returnedListAll[1]).toMatchObject(xmlApiData.roListAll[1])
-		expect(returnedListAll).toMatchSnapshot()
+		await mos.dispose()
 	})
 })
