--- src/__tests__/index.spec.ts
@@ -1,4 +1,0 @@
-
-test('Simple test', () => {
-	expect(2).toBe(2)
-})
